Re-syncing from source repos.
That is git@github.com:konradwilk/fibonacci.git
and git@github.com:konradwilk/sha1.git both have

LICENSE: Sync up all the files with the top level license

They are all under the same license, but make it more
explicit by having it in of the source code.

And this follows the same change.

Signed-off-by: Konrad Rzeszutek Wilk <konrad@kernel.org>